requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-community>=0.0.10
requests>=2.31.0
sqlite3
python-dotenv>=1.0.0